requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
flask
flask-cors
requests